requirements.txt
aiohttp >= 1.0.5
aiohttp-jinja2 >= 0.8.0
Jinja2 >= 2.8
numpy >= 1.11.1
pandas >= 0.18.1
Pillow >= 3.3.1
pyzmq >= 15.4.0
scipy >= 0.18.1
